feat: full raytracing demo

+++ PyriteCore/src/scene/scene.h
@@ -8,14 +8,14 @@ class Scene
 public:
   virtual ~Scene() = default;
   
-  virtual void update(double delta) = 0;
+  virtual void update(float delta) = 0;
   virtual void render() = 0;
 };
 
 class EmptyScene : public Scene
 {
 public:
-  void update(double delta) override {}
+  void update(float delta) override {}
   void render() override {}
 };
